Move GUI components to application source tree. #3
