+++ .gitmodules
@@ -8,7 +8,7 @@
 	path = x86_64/lv_drivers
 	url = https://github.com/lvgl/lv_drivers.git
 [submodule "lib/gui/lv_lib_qrcode"]
-	path = esp32/components/gui/lv_lib_qrcode
+	path = esp32/main/lib/gui/lv_lib_qrcode
 	url = https://github.com/lvgl/lv_lib_qrcode.git
 [submodule "esp32/components/esp32-button"]
 	path = esp32/components/esp32-button
