+++ esp32/main/lib/gui/assets/black_60.c
