+++ esp32/main/lib/gui/assets/btc-icon-60.png
