+++ esp32/main/lib/gui/assets/btc_icon_60.c
