+++ esp32/main/lib/gui/assets/dge-icon-60.png
