+++ esp32/main/lib/gui/assets/dge_icon_60.c
