+++ esp32/main/lib/gui/assets/eth-icon-60.png
