+++ esp32/main/lib/gui/assets/eth_icon_60.c
