+++ esp32/main/lib/gui/assets/ltc_icon_60.c
