+++ esp32/main/lib/gui/assets/warning-30.png
