+++ esp32/main/lib/gui/assets/warning_30.c
