+++ esp32/main/lib/gui/tools/format_percentage.hpp
