+++ esp32/main/lib/gui/widgets/composite/quote_banner.hpp
