+++ esp32/main/lib/gui/widgets/elemental/image.hpp
