+++ esp32/main/lib/gui/widgets/elemental/qrcode.hpp
