+++ x86_64/main.cpp
@@ -2,10 +2,10 @@
 #include <unistd.h>
 
 #include <cstdio>
-#include <gui/hal/driver.hpp>
+#include <hal/lvgl_driver.hpp>
 
 int main(int, char**) {
-  GUI::HAL::LVGL()->init();
+  HAL::LVGL()->init();
   while(1) {
     usleep(1000 * 100);
   }
