+++ x86_64/provisioning_screen.cpp
@@ -2,11 +2,11 @@
 #include <unistd.h>
 
 #include <cstdio>
-#include <gui/hal/driver.hpp>
+#include <hal/lvgl_driver.hpp>
 #include <gui/views/startup/provisioning.hpp>
 
 int main(int, char**) {
-  GUI::HAL::LVGL()->init();
+  HAL::LVGL()->init();
   GUI::ProvisioningScreen()->setQR("{\"ver\":\"v1\",\"name\":\"PROV_ticker\",\"pop\":\"abcd1234\",\"transport\":\"softap\"}");
   GUI::ProvisioningScreen()->show();
